requirements.txt
backtrader==1.9.59.122
certifi==2018.1.18
chardet==3.0.4
click==6.7
cssselect==1.0.3
cycler==0.10.0
idna==2.6
Keras==2.1.3
lxml==4.1.1
matplotlib==2.1.2
nose==1.3.7
numpy==1.14.0
pycrypto==2.6.1
pyparsing==2.2.0
pyquery==1.4.0
python-dateutil==2.6.1
python-dotenv==0.7.1
pytz==2017.3
PyYAML==3.12
requests==2.18.4
scipy==1.0.0
six==1.11.0
tornado==4.5.3
urllib3==1.22
